+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
-## v3.??.?? (2024-02-13)
+## v3.66.0 (2024-02-13)
 
 #### :boom: Breaking Change